+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/ui/player/EffectsViewModel.kt
@@ -2,7 +2,7 @@ package au.com.shiftyjelly.pocketcasts.wear.ui.player
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import au.com.shiftyjelly.pocketcasts.models.to.PlaybackEffects
+import au.com.shiftyjelly.pocketcasts.models.to.PlaybackEffectsData
 import au.com.shiftyjelly.pocketcasts.models.type.TrimMode
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
@@ -13,7 +13,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.rx2.asFlow
 import javax.inject.Inject
 
 @HiltViewModel
@@ -24,11 +23,8 @@ class EffectsViewModel
 ) : ViewModel() {
 
     val state: StateFlow<State> =
-        playbackManager.playbackStateRelay
-            .asFlow()
-            .map {
-                State.Loaded(settings.globalPlaybackEffects.value)
-            }
+        settings.globalPlaybackEffects.flow
+            .map { State.Loaded(it.toData()) }
             .stateIn(
                 scope = viewModelScope,
                 started = SharingStarted.WhileSubscribed(stopTimeoutMillis = 5_000),
@@ -48,25 +44,26 @@ class EffectsViewModel
     private fun changePlaybackSpeed(speed: Double) {
         val currentState = state.value as? State.Loaded ?: return
         val effects = currentState.playbackEffects
-        effects.playbackSpeed = speed.roundedSpeed()
-        saveEffects(effects)
+        val newEffects = effects.copy(playbackSpeed = speed.roundedSpeed())
+        saveEffects(newEffects)
     }
 
     fun updateTrimSilence(trimMode: TrimMode) {
         val currentState = state.value as? State.Loaded ?: return
         val effects = currentState.playbackEffects
-        effects.trimMode = trimMode
-        saveEffects(effects)
+        val newEffects = effects.copy(trimMode = trimMode)
+        saveEffects(newEffects)
     }
 
     fun updateBoostVolume(boostVolume: Boolean) {
         val currentState = state.value as? State.Loaded ?: return
         val effects = currentState.playbackEffects
-        effects.isVolumeBoosted = boostVolume
-        saveEffects(effects)
+        val newEffects = effects.copy(isVolumeBoosted = boostVolume)
+        saveEffects(newEffects)
     }
 
-    private fun saveEffects(effects: PlaybackEffects) {
+    private fun saveEffects(effectsData: PlaybackEffectsData) {
+        val effects = effectsData.toEffects()
         viewModelScope.launch {
             playbackManager.updatePlayerEffects(effects)
             settings.globalPlaybackEffects.set(effects)
@@ -75,7 +72,7 @@ class EffectsViewModel
 
     sealed class State {
         data class Loaded(
-            val playbackEffects: PlaybackEffects,
+            val playbackEffects: PlaybackEffectsData,
         ) : State()
 
         object Loading : State()